Create Temperature Converter Website
Developing a website that converts temperature recorded in one unit to
another can be an excellent place for web developers to move forward in their web development journey. The measuring units of the temperature recorded in
a particular unit can be converted with a temperature converter, necessitating
you to build a dropdown menu with temperature scales.