Add files via upload
Background Generator
A background generator is a great way to practice CSS skills and
familiarize yourself with basic JavaScript concepts. In this project, you will select a basic or a gradient colour and generate
it via code. You will then create a webpage that generates
random background colours and allows users to customize and
copy the generated colour code.